[refac]: renamed file "abstract.py" to "oddelement.py"
