+++ tests/utils/docs/test_odd2md.py
@@ -8,10 +8,10 @@ from utils.docs.odd2md import (
     ODD2Md,
     ODDReader,
 )
-from utils.docs.specs.abstract import ODDElement
 from utils.docs.specs.attributespec import AttributeSpec
 from utils.docs.specs.elementspec import ElementSpec
 from utils.docs.specs.namespaces import NS_MAP
+from utils.docs.specs.oddelement import ODDElement
 from utils.docs.specs.utils import split_tag_and_ns
 
 from .conftest import EL_FINDER, check_result_with_xpath
